+++ lib/rails_edge_test/dsl/edge.rb
@@ -38,6 +38,18 @@ module RailsEdgeTest::Dsl
       process(parameters)
     end
 
+    def perform_delete(parameters={})
+      request.instance_variable_set(:@method, "DELETE")
+      request.env['REQUEST_METHOD'] = "DELETE"
+      process(parameters)
+    end
+
+    def set_authenticity_token
+      r = request
+      controller.instance_eval { @_request = r }
+      request.headers['X-CSRF-Token'] = controller.send :form_authenticity_token
+    end
+
     def process(parameters={})
       request.assign_parameters(
         ::Rails.application.routes,